backend/requirements.txt
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# HTTP Client
requests==2.31.0

# Data Processing
numpy==1.24.3

# Audio Processing
librosa==0.10.1
soundfile==0.12.1

# Machine Learning
tensorflow==2.15.0
scikit-learn==1.3.2

# External APIs
deepgram-sdk==3.2.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.12.0
flake8==6.1.0
mypy==1.7.1

